wrap up, part 1
